requirements.txt
python-telegram-bot
requests
sqlalchemy==2.0.18
python-dotenv>=1.0.0
dextools-python
selenium==4.17.2

python-coinmarketcap
mplfinance
dexscreener

psycopg2
nowpayment